backend/requirements.txt
boto3>=1.34.0
aws-lambda-powertools[tracer]>=2.32.0
pyjwt>=2.8.0
aws-xray-sdk>=2.12.0
